+++ src/prisma-generator.ts
@@ -6,13 +6,9 @@ import {
 } from '@prisma/generator-helper';
 import { getDMMF, parseEnvValue } from '@prisma/internals';
 import { promises as fs } from 'fs';
-import { isAggregateOutputType } from './helpers';
 import Transformer from './transformer';
 import removeDir from './utils/removeDir';
-import {
-  addMissingInputObjectTypes,
-  handleMongoDbRawOperationsAndQueries,
-} from './helpers';
+import { addMissingInputObjectTypes } from './helpers';
 
 export async function generate(options: GeneratorOptions) {
   await handleGeneratorOutputValue(options.generator.output as EnvValue);
@@ -42,23 +38,20 @@ export async function generate(options: GeneratorOptions) {
   const dataSource = options.datasources?.[0];
   Transformer.provider = dataSource.provider;
 
-  // TODO: remove once Prisma fix this issue: https://github.com/prisma/prisma/issues/14900
-  if (dataSource.provider === 'mongodb') {
-    handleMongoDbRawOperationsAndQueries(
-      modelOperations,
-      outputObjectTypes,
-      inputObjectTypes,
-    );
-  }
-
-  addMissingInputObjectTypes(inputObjectTypes, outputObjectTypes, models);
+  addMissingInputObjectTypes(
+    inputObjectTypes,
+    outputObjectTypes,
+    models,
+    modelOperations,
+    dataSource.provider,
+  );
   await generateObjectSchemas(inputObjectTypes);
 
   await generateModelSchemas(modelOperations);
 }
 
-async function handleGeneratorOutputValue(generatorOuputValue: EnvValue) {
-  const outputDirectoryPath = parseEnvValue(generatorOuputValue);
+async function handleGeneratorOutputValue(generatorOutputValue: EnvValue) {
+  const outputDirectoryPath = parseEnvValue(generatorOutputValue);
 
   // create the output directory and delete contents that might exist from a previous run
   await fs.mkdir(outputDirectoryPath, { recursive: true });
@@ -81,39 +74,6 @@ function checkForCustomPrismaClientOutputPath(
   if (prismaClientGeneratorConfig?.isCustomOutput) {
     Transformer.setPrismaClientOutputPath(
       prismaClientGeneratorConfig.output?.value as string,
-  const inputObjectTypes = prismaClientDmmf.schema.inputObjectTypes.prisma;
-  const outputObjectTypes = prismaClientDmmf.schema.outputObjectTypes.prisma;
-
-  outputObjectTypes.forEach((outputObjectType) => {
-    if (isAggregateOutputType(outputObjectType.name)) {
-      const name = outputObjectType.name.replace(/(?:OutputType|Output)$/, '');
-      inputObjectTypes.push({
-        constraints: { maxNumFields: null, minNumFields: null },
-        name: `${name}Input`,
-        fields: outputObjectType.fields.map((field) => ({
-          name: field.name,
-          isNullable: false,
-          isRequired: false,
-          inputTypes: [
-            {
-              isList: false,
-              type: 'True',
-              location: 'scalar',
-            },
-          ],
-        })),
-      });
-    }
-  });
-
-  const rawOpsMap: { [name: string]: string } = {};
-
-  /*
-  TODO: remove once Prisma fix this issue: https://github.com/prisma/prisma/issues/14900
-  */
-  if (dataSource.provider === 'mongodb') {
-    const modelNames = prismaClientDmmf.mappings.modelOperations.map(
-      (item) => item.model,
     );
   }
 }
